Handle banded fee calculation for 'unlimited' sum
Handle the case when the amount involved in banded fee calculation
exceeds the upper limit. Currently the upper limit it's being stored as
just a string 'unlimited'.

The rest of the values are all stored as integers, instead of floats, as
it seems easier at this stage. One of the main reasons is that the
'unlimited' upper band needs to be held too, but can't think of a good
way to do it.